Undo the changes to flags.patch
The update to flags.patch in the last patch was incorrect. Instead of
setting the extra compiler flags to the libsqlite3 library, it set the
flags to the sqlite3 binary. That resulted in functions missing from
libsqlite3, which in turn caused the QtSql plugin for sqlite3 to fail to
load and qtbase to be unable to compile.